+++ scripts/docker/docker-entrypoint.sh
@@ -49,18 +49,14 @@ if [ ! -f /var/run/gerbera/config.xml ]; then
   fi
 
   # Generate a config file with home set
-  gerbera --create-config --home /var/run/gerbera > /var/run/gerbera/config.xml
+  gerbera --create-config --home /var/run/gerbera --scripts /mnt/customization/js --modules=Autoscan > /var/run/gerbera/config.xml
 
   # Automatically scan /content with inotify (for a volume mount)
-  sed 's|<import hidden-files="no">|<import hidden-files="no">\n\
+  echo '\
     <autoscan use-inotify="yes">\n\
-    <directory location="/mnt/content" mode="inotify" \
-               recursive="yes" hidden-files="no"/>\n\
-    </autoscan>|' -i /var/run/gerbera/config.xml
-
-  # Add directory for custom JavaScript scripts
-  sed 's|</common>|</common>\
-        <custom>/mnt/customization/js</custom>|' -i /var/run/gerbera/config.xml
+      <directory location="/mnt/content" mode="inotify" \
+                 recursive="yes" hidden-files="no"/>\n\
+    </autoscan>' > /var/run/gerbera/autoscan.xml
 
   # Allow customization of Gerbera configuration file
   if [ -x /mnt/customization/shell/gerbera_config.sh ]; then