+++ src/config/config_manager.cc
@@ -111,6 +111,42 @@ void ConfigManager::addOption(ConfigVal option, const std::shared_ptr<ConfigOpti
     options.at(to_underlying(option)) = optionValue;
 }
 
+void ConfigManager::expandFiles(pugi::xml_node& parent)
+{
+    auto cs = definition->removeAttribute(ConfigVal::A_LOAD_SECTION_FROM_FILE);
+    std::vector<pugi::xml_node> nodesToRemove;
+    for (auto&& node : parent.children()) {
+        if (node.attribute(cs.c_str())) {
+            std::error_code ec;
+            pugi::xml_document xmlSubDoc;
+            fs::path subModuleName = this->filename.parent_path() / node.attribute(cs.c_str()).value();
+            log_info("Merging config file '{}'", subModuleName.c_str());
+            fs::directory_entry dirEnt(subModuleName, ec);
+
+            if (!isRegularFile(dirEnt, ec) && !dirEnt.is_symlink(ec)) {
+                log_warning("Configuration sub module '{}' does not exist for '{}'", subModuleName.c_str(), node.name());
+                expandFiles(node);
+            } else {
+                pugi::xml_parse_result result = xmlSubDoc.load_file(subModuleName.c_str());
+                if (result.status != pugi::xml_parse_status::status_ok) {
+                    throw ConfigParseException(result.description());
+                }
+                log_info("Parsing configuration {} ...", subModuleName.c_str());
+
+                auto subRoot = xmlSubDoc.document_element();
+                subRoot = parent.insert_copy_before(subRoot, node);
+                nodesToRemove.push_back(node);
+                expandFiles(subRoot);
+            }
+        } else if (node.type() == pugi::xml_node_type::node_element) {
+            expandFiles(node);
+        }
+    }
+    for (auto&& node : nodesToRemove) {
+        parent.remove_child(node);
+    }
+}
+
 void ConfigManager::getAllNodes(const pugi::xml_node& parent, bool getAttributes)
 {
     for (auto&& node : parent.children()) {
@@ -146,6 +182,7 @@ void ConfigManager::load(const fs::path& userHome)
 
     auto root = xmlDoc.document_element();
 
+    expandFiles(root);
     getAllNodes(root, false);
 
     // first check if the config file itself looks ok,